+++ tests/testthat/test-read_redcap_tables.R
@@ -1,4 +1,3 @@
-# Unit Test
 
 # Test that the function throws an error when uri and token are not provided
 test_that("read_redcap_tables throws error when uri and token are not provided",